Added Gradle wrapper to be able to build the project out of the box.
To not alienate users not familiar with Gradle and also make sure that a relatively recent version of Gradle is used to build the project, this commit adds the Gradle wrapper task. This allows users to build the project out of the box without installation of any further software.

Added build information to the readme.